+++ src/utils/utils.c
@@ -154,7 +154,7 @@ void parseArgs(int argc, char* const argv[]) {
               break;
 
           case 'm':
-              markerDelay = atoi(optarg);
+              markerDelay = atof(optarg);
               break;
 
           case 's':
@@ -178,7 +178,7 @@ void parseArgs(int argc, char* const argv[]) {
   debug("Arg network set to %s", network);
   debug("Arg inputHostname set to %s", inputHostname);
   debug("Arg hostsFilePath set to %s", hostsFilePath);
-  debug("Arg markerDelay set to %d", markerDelay);
+  debug("Arg markerDelay set to %0.2f", markerDelay);
   debug("Arg snapshotDelay set to %d", snapshotDelay);
   debug("Arg tokenDelay set to %0.2f", tokenDelay);
   debug("Arg snapshotId set to %d", snapshotId);
